run duplicate row matcher in both directions, see #27
Duplicate rows are difficult for daff to do something sensible
with, as its matcher is currently optimized for tables with
a set of columns that approximate a primary key.  It does have
a heuristic to make obvious matches for identical rows that lie
beside matched rows.  Previously this heuristic was asymmetric,
this commit fixes that.