update python2 fixes for latest haxe
Haxe doesn't support python2, so we use the 3to2 utility to make
haxe-generated python usable in python2.  Then we patch things
up to make the same code still runnable in python3.  This is getting
a little creaky, and it will soon be preferable to publish a wheel
with multiple distinct code versions for different python flavors.