add support for psv
This adds support for the psv format.  This is not easy in
javascript due to the poop symbol lying on an astral plane,
so for now I've focused on getting it right in python.

Now allow `daff foo.bar` to be used instead of `daff copy foo.bar`,
for convenience.

Small cleanup to the python install rat's-nest.