prepare to use node:sqlite (#208)
The `fiber` dependency in daff.js has always been rather silly. At the time I wrote the node code, `node-sqlite3` was the "standard" for working with sqlite3 in node, but it was async, and hard to integrate with translated code. Now, node has built-in support for sqlite3 on the way, and also `better-sqlite3` is a find sync interface.

Here I rework the node helper files to use `node:sqlite` if available, and `better-sqlite3` if not. I've removed all node dependencies, which I'm not very happy about, but daff sees a lot of use without sqlite3 so it seems bad to be forcing people to install dependencies that include compiled native code.